Manage connection instance locally
Since the connection was kept in class variables, using this gem for
multiple prestashop ecommerces would not be possible without causing credential conflicts.
The chosen solution was to hold the connection instance locally instead
of a singleton class. The gem structure was preserved, although I
changed the client class name to make the incompatibility obvious.